LayoutTests/svg/custom/global-constructors-expected.txt

Test to make sure we expose all the global constructor objects requested by http://www.w3.org/TR/SVG11/ecmascript-binding.html

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS SVGException.toString() is '[object SVGExceptionConstructor]'
PASS SVGElement.toString() is '[object SVGElementConstructor]'
PASS SVGAnimatedBoolean.toString() is '[object SVGAnimatedBooleanConstructor]'
PASS SVGAnimatedString.toString() is '[object SVGAnimatedStringConstructor]'
PASS SVGStringList.toString() is '[object SVGStringListConstructor]'
PASS SVGAnimatedEnumeration.toString() is '[object SVGAnimatedEnumerationConstructor]'
PASS SVGAnimatedInteger.toString() is '[object SVGAnimatedIntegerConstructor]'
PASS SVGNumber.toString() is '[object SVGNumberConstructor]'
PASS SVGAnimatedNumber.toString() is '[object SVGAnimatedNumberConstructor]'
PASS SVGNumberList.toString() is '[object SVGNumberListConstructor]'
PASS SVGAnimatedNumberList.toString() is '[object SVGAnimatedNumberListConstructor]'
PASS SVGLength.toString() is '[object SVGLengthConstructor]'
PASS SVGAnimatedLength.toString() is '[object SVGAnimatedLengthConstructor]'
PASS SVGLengthList.toString() is '[object SVGLengthListConstructor]'
PASS SVGAnimatedLengthList.toString() is '[object SVGAnimatedLengthListConstructor]'
PASS SVGAngle.toString() is '[object SVGAngleConstructor]'
PASS SVGAnimatedAngle.toString() is '[object SVGAnimatedAngleConstructor]'
PASS SVGColor.toString() is '[object SVGColorConstructor]'
FAIL SVGICCColor.toString() should be [object SVGICCColorConstructor]. Threw exception ReferenceError: Can't find variable: SVGICCColor
PASS SVGRect.toString() is '[object SVGRectConstructor]'
PASS SVGAnimatedRect.toString() is '[object SVGAnimatedRectConstructor]'
FAIL SVGStylable.toString() should be [object SVGStylableConstructor]. Threw exception ReferenceError: Can't find variable: SVGStylable
FAIL SVGLocatable.toString() should be [object SVGLocatableConstructor]. Threw exception ReferenceError: Can't find variable: SVGLocatable
FAIL SVGTransformable.toString() should be [object SVGTransformableConstructor]. Threw exception ReferenceError: Can't find variable: SVGTransformable
FAIL SVGTests.toString() should be [object SVGTestsConstructor]. Threw exception ReferenceError: Can't find variable: SVGTests
FAIL SVGLangSpace.toString() should be [object SVGLangSpaceConstructor]. Threw exception ReferenceError: Can't find variable: SVGLangSpace
FAIL SVGExternalResourcesRequired.toString() should be [object SVGExternalResourcesRequiredConstructor]. Threw exception ReferenceError: Can't find variable: SVGExternalResourcesRequired
FAIL SVGFitToViewBox.toString() should be [object SVGFitToViewBoxConstructor]. Threw exception ReferenceError: Can't find variable: SVGFitToViewBox
FAIL SVGZoomAndPan.toString() should be [object SVGZoomAndPanConstructor]. Threw exception ReferenceError: Can't find variable: SVGZoomAndPan
FAIL SVGViewSpec.toString() should be [object SVGViewSpecConstructor]. Threw exception ReferenceError: Can't find variable: SVGViewSpec
FAIL SVGURIReference.toString() should be [object SVGURIReferenceConstructor]. Threw exception ReferenceError: Can't find variable: SVGURIReference
FAIL SVGCSSRule.toString() should be [object SVGCSSRuleConstructor]. Threw exception ReferenceError: Can't find variable: SVGCSSRule
PASS SVGDocument.toString() is '[object SVGDocumentConstructor]'
PASS SVGSVGElement.toString() is '[object SVGSVGElementConstructor]'
PASS SVGGElement.toString() is '[object SVGGElementConstructor]'
PASS SVGDefsElement.toString() is '[object SVGDefsElementConstructor]'
PASS SVGDescElement.toString() is '[object SVGDescElementConstructor]'
PASS SVGTitleElement.toString() is '[object SVGTitleElementConstructor]'
PASS SVGSymbolElement.toString() is '[object SVGSymbolElementConstructor]'
PASS SVGUseElement.toString() is '[object SVGUseElementConstructor]'
PASS SVGElementInstance.toString() is '[object SVGElementInstanceConstructor]'
PASS SVGElementInstanceList.toString() is '[object SVGElementInstanceListConstructor]'
PASS SVGImageElement.toString() is '[object SVGImageElementConstructor]'
PASS SVGSwitchElement.toString() is '[object SVGSwitchElementConstructor]'
PASS SVGStyleElement.toString() is '[object SVGStyleElementConstructor]'
PASS SVGPoint.toString() is '[object SVGPointConstructor]'
PASS SVGPointList.toString() is '[object SVGPointListConstructor]'
PASS SVGMatrix.toString() is '[object SVGMatrixConstructor]'
PASS SVGTransform.toString() is '[object SVGTransformConstructor]'
PASS SVGTransformList.toString() is '[object SVGTransformListConstructor]'
PASS SVGAnimatedTransformList.toString() is '[object SVGAnimatedTransformListConstructor]'
PASS SVGPreserveAspectRatio.toString() is '[object SVGPreserveAspectRatioConstructor]'
PASS SVGAnimatedPreserveAspectRatio.toString() is '[object SVGAnimatedPreserveAspectRatioConstructor]'
PASS SVGPathSeg.toString() is '[object SVGPathSegConstructor]'
PASS SVGPathSegClosePath.toString() is '[object SVGPathSegClosePathConstructor]'
PASS SVGPathSegMovetoAbs.toString() is '[object SVGPathSegMovetoAbsConstructor]'
PASS SVGPathSegMovetoRel.toString() is '[object SVGPathSegMovetoRelConstructor]'
PASS SVGPathSegLinetoAbs.toString() is '[object SVGPathSegLinetoAbsConstructor]'
PASS SVGPathSegLinetoRel.toString() is '[object SVGPathSegLinetoRelConstructor]'
PASS SVGPathSegCurvetoCubicAbs.toString() is '[object SVGPathSegCurvetoCubicAbsConstructor]'
PASS SVGPathSegCurvetoCubicRel.toString() is '[object SVGPathSegCurvetoCubicRelConstructor]'
PASS SVGPathSegCurvetoQuadraticAbs.toString() is '[object SVGPathSegCurvetoQuadraticAbsConstructor]'
PASS SVGPathSegCurvetoQuadraticRel.toString() is '[object SVGPathSegCurvetoQuadraticRelConstructor]'
PASS SVGPathSegArcAbs.toString() is '[object SVGPathSegArcAbsConstructor]'
PASS SVGPathSegArcRel.toString() is '[object SVGPathSegArcRelConstructor]'
PASS SVGPathSegLinetoHorizontalAbs.toString() is '[object SVGPathSegLinetoHorizontalAbsConstructor]'
PASS SVGPathSegLinetoHorizontalRel.toString() is '[object SVGPathSegLinetoHorizontalRelConstructor]'
PASS SVGPathSegLinetoVerticalAbs.toString() is '[object SVGPathSegLinetoVerticalAbsConstructor]'
PASS SVGPathSegLinetoVerticalRel.toString() is '[object SVGPathSegLinetoVerticalRelConstructor]'
PASS SVGPathSegCurvetoCubicSmoothAbs.toString() is '[object SVGPathSegCurvetoCubicSmoothAbsConstructor]'
PASS SVGPathSegCurvetoCubicSmoothRel.toString() is '[object SVGPathSegCurvetoCubicSmoothRelConstructor]'
PASS SVGPathSegCurvetoQuadraticSmoothAbs.toString() is '[object SVGPathSegCurvetoQuadraticSmoothAbsConstructor]'
PASS SVGPathSegCurvetoQuadraticSmoothRel.toString() is '[object SVGPathSegCurvetoQuadraticSmoothRelConstructor]'
PASS SVGPathSegList.toString() is '[object SVGPathSegListConstructor]'
FAIL SVGAnimatedPathData.toString() should be [object SVGAnimatedPathDataConstructor]. Threw exception ReferenceError: Can't find variable: SVGAnimatedPathData
PASS SVGPathElement.toString() is '[object SVGPathElementConstructor]'
PASS SVGRectElement.toString() is '[object SVGRectElementConstructor]'
PASS SVGCircleElement.toString() is '[object SVGCircleElementConstructor]'
PASS SVGEllipseElement.toString() is '[object SVGEllipseElementConstructor]'
PASS SVGLineElement.toString() is '[object SVGLineElementConstructor]'
FAIL SVGAnimatedPoints.toString() should be [object SVGAnimatedPointsConstructor]. Threw exception ReferenceError: Can't find variable: SVGAnimatedPoints
PASS SVGPolylineElement.toString() is '[object SVGPolylineElementConstructor]'
PASS SVGPolygonElement.toString() is '[object SVGPolygonElementConstructor]'
PASS SVGTextContentElement.toString() is '[object SVGTextContentElementConstructor]'
PASS SVGTextPositioningElement.toString() is '[object SVGTextPositioningElementConstructor]'
PASS SVGTextElement.toString() is '[object SVGTextElementConstructor]'
PASS SVGTSpanElement.toString() is '[object SVGTSpanElementConstructor]'
PASS SVGTRefElement.toString() is '[object SVGTRefElementConstructor]'
PASS SVGTextPathElement.toString() is '[object SVGTextPathElementConstructor]'
PASS SVGAltGlyphElement.toString() is '[object SVGAltGlyphElementConstructor]'
FAIL SVGAltGlyphDefElement.toString() should be [object SVGAltGlyphDefElementConstructor]. Threw exception ReferenceError: Can't find variable: SVGAltGlyphDefElement
FAIL SVGAltGlyphItemElement.toString() should be [object SVGAltGlyphItemElementConstructor]. Threw exception ReferenceError: Can't find variable: SVGAltGlyphItemElement
FAIL SVGGlyphRefElement.toString() should be [object SVGGlyphRefElementConstructor]. Threw exception ReferenceError: Can't find variable: SVGGlyphRefElement
PASS SVGPaint.toString() is '[object SVGPaintConstructor]'
PASS SVGMarkerElement.toString() is '[object SVGMarkerElementConstructor]'
FAIL SVGColorProfileElement.toString() should be [object SVGColorProfileElementConstructor]. Threw exception ReferenceError: Can't find variable: SVGColorProfileElement
FAIL SVGColorProfileRule.toString() should be [object SVGColorProfileRuleConstructor]. Threw exception ReferenceError: Can't find variable: SVGColorProfileRule
PASS SVGGradientElement.toString() is '[object SVGGradientElementConstructor]'
PASS SVGLinearGradientElement.toString() is '[object SVGLinearGradientElementConstructor]'
PASS SVGRadialGradientElement.toString() is '[object SVGRadialGradientElementConstructor]'
PASS SVGStopElement.toString() is '[object SVGStopElementConstructor]'
PASS SVGPatternElement.toString() is '[object SVGPatternElementConstructor]'
PASS SVGClipPathElement.toString() is '[object SVGClipPathElementConstructor]'
PASS SVGMaskElement.toString() is '[object SVGMaskElementConstructor]'
FAIL SVGFilterElement.toString() should be [object SVGFilterElementConstructor]. Threw exception ReferenceError: Can't find variable: SVGFilterElement
FAIL SVGFilterPrimitiveStandardAttributes.toString() should be [object SVGFilterPrimitiveStandardAttributesConstructor]. Threw exception ReferenceError: Can't find variable: SVGFilterPrimitiveStandardAttributes
FAIL SVGFEBlendElement.toString() should be [object SVGFEBlendElementConstructor]. Threw exception ReferenceError: Can't find variable: SVGFEBlendElement
FAIL SVGFEColorMatrixElement.toString() should be [object SVGFEColorMatrixElementConstructor]. Threw exception ReferenceError: Can't find variable: SVGFEColorMatrixElement
FAIL SVGFEComponentTransferElement.toString() should be [object SVGFEComponentTransferElementConstructor]. Threw exception ReferenceError: Can't find variable: SVGFEComponentTransferElement
FAIL SVGComponentTransferFunctionElement.toString() should be [object SVGComponentTransferFunctionElementConstructor]. Threw exception ReferenceError: Can't find variable: SVGComponentTransferFunctionElement
FAIL SVGFEFuncRElement.toString() should be [object SVGFEFuncRElementConstructor]. Threw exception ReferenceError: Can't find variable: SVGFEFuncRElement
FAIL SVGFEFuncGElement.toString() should be [object SVGFEFuncGElementConstructor]. Threw exception ReferenceError: Can't find variable: SVGFEFuncGElement
FAIL SVGFEFuncBElement.toString() should be [object SVGFEFuncBElementConstructor]. Threw exception ReferenceError: Can't find variable: SVGFEFuncBElement
FAIL SVGFEFuncAElement.toString() should be [object SVGFEFuncAElementConstructor]. Threw exception ReferenceError: Can't find variable: SVGFEFuncAElement
FAIL SVGFECompositeElement.toString() should be [object SVGFECompositeElementConstructor]. Threw exception ReferenceError: Can't find variable: SVGFECompositeElement
FAIL SVGFEConvolveMatrixElement.toString() should be [object SVGFEConvolveMatrixElementConstructor]. Threw exception ReferenceError: Can't find variable: SVGFEConvolveMatrixElement
FAIL SVGFEDiffuseLightingElement.toString() should be [object SVGFEDiffuseLightingElementConstructor]. Threw exception ReferenceError: Can't find variable: SVGFEDiffuseLightingElement
FAIL SVGFEDistantLightElement.toString() should be [object SVGFEDistantLightElementConstructor]. Threw exception ReferenceError: Can't find variable: SVGFEDistantLightElement
FAIL SVGFEPointLightElement.toString() should be [object SVGFEPointLightElementConstructor]. Threw exception ReferenceError: Can't find variable: SVGFEPointLightElement
FAIL SVGFESpotLightElement.toString() should be [object SVGFESpotLightElementConstructor]. Threw exception ReferenceError: Can't find variable: SVGFESpotLightElement
FAIL SVGFEDisplacementMapElement.toString() should be [object SVGFEDisplacementMapElementConstructor]. Threw exception ReferenceError: Can't find variable: SVGFEDisplacementMapElement
FAIL SVGFEFloodElement.toString() should be [object SVGFEFloodElementConstructor]. Threw exception ReferenceError: Can't find variable: SVGFEFloodElement
FAIL SVGFEGaussianBlurElement.toString() should be [object SVGFEGaussianBlurElementConstructor]. Threw exception ReferenceError: Can't find variable: SVGFEGaussianBlurElement
FAIL SVGFEImageElement.toString() should be [object SVGFEImageElementConstructor]. Threw exception ReferenceError: Can't find variable: SVGFEImageElement
FAIL SVGFEMergeElement.toString() should be [object SVGFEMergeElementConstructor]. Threw exception ReferenceError: Can't find variable: SVGFEMergeElement
FAIL SVGFEMergeNodeElement.toString() should be [object SVGFEMergeNodeElementConstructor]. Threw exception ReferenceError: Can't find variable: SVGFEMergeNodeElement
FAIL SVGFEMorphologyElement.toString() should be [object SVGFEMorphologyElementConstructor]. Threw exception ReferenceError: Can't find variable: SVGFEMorphologyElement
FAIL SVGFEOffsetElement.toString() should be [object SVGFEOffsetElementConstructor]. Threw exception ReferenceError: Can't find variable: SVGFEOffsetElement
FAIL SVGFESpecularLightingElement.toString() should be [object SVGFESpecularLightingElementConstructor]. Threw exception ReferenceError: Can't find variable: SVGFESpecularLightingElement
FAIL SVGFETileElement.toString() should be [object SVGFETileElementConstructor]. Threw exception ReferenceError: Can't find variable: SVGFETileElement
FAIL SVGFETurbulenceElement.toString() should be [object SVGFETurbulenceElementConstructor]. Threw exception ReferenceError: Can't find variable: SVGFETurbulenceElement
PASS SVGCursorElement.toString() is '[object SVGCursorElementConstructor]'
PASS SVGAElement.toString() is '[object SVGAElementConstructor]'
PASS SVGViewElement.toString() is '[object SVGViewElementConstructor]'
PASS SVGScriptElement.toString() is '[object SVGScriptElementConstructor]'
FAIL SVGEvent.toString() should be [object SVGEventConstructor]. Threw exception ReferenceError: Can't find variable: SVGEvent
PASS SVGZoomEvent.toString() is '[object SVGZoomEventConstructor]'
FAIL SVGAnimationElement.toString() should be [object SVGAnimationElementConstructor]. Threw exception ReferenceError: Can't find variable: SVGAnimationElement
PASS SVGAnimateElement.toString() is '[object SVGAnimateElementConstructor]'
PASS SVGSetElement.toString() is '[object SVGSetElementConstructor]'
FAIL SVGAnimateMotionElement.toString() should be [object SVGAnimateMotionElementConstructor]. Threw exception ReferenceError: Can't find variable: SVGAnimateMotionElement
FAIL SVGMPathElement.toString() should be [object SVGMPathElementConstructor]. Threw exception ReferenceError: Can't find variable: SVGMPathElement
PASS SVGAnimateColorElement.toString() is '[object SVGAnimateColorElementConstructor]'
PASS SVGAnimateTransformElement.toString() is '[object SVGAnimateTransformElementConstructor]'
PASS SVGFontElement.toString() is '[object SVGFontElementConstructor]'
PASS SVGGlyphElement.toString() is '[object SVGGlyphElementConstructor]'
PASS SVGMissingGlyphElement.toString() is '[object SVGMissingGlyphElementConstructor]'
PASS SVGHKernElement.toString() is '[object SVGHKernElementConstructor]'
PASS SVGVKernElement.toString() is '[object SVGVKernElementConstructor]'
PASS SVGFontFaceElement.toString() is '[object SVGFontFaceElementConstructor]'
PASS SVGFontFaceSrcElement.toString() is '[object SVGFontFaceSrcElementConstructor]'
PASS SVGFontFaceUriElement.toString() is '[object SVGFontFaceUriElementConstructor]'
PASS SVGFontFaceFormatElement.toString() is '[object SVGFontFaceFormatElementConstructor]'
PASS SVGFontFaceNameElement.toString() is '[object SVGFontFaceNameElementConstructor]'
FAIL SVGDefinitionSrcElement.toString() should be [object SVGDefinitionSrcElementConstructor]. Threw exception ReferenceError: Can't find variable: SVGDefinitionSrcElement
PASS SVGMetadataElement.toString() is '[object SVGMetadataElementConstructor]'
PASS SVGForeignObjectElement.toString() is '[object SVGForeignObjectElementConstructor]'
PASS successfullyParsed is true

TEST COMPLETE
